Moved treesitter functions to separate file
